gopls: Measure the efficacy of completions
This CL introduces code to count how often proposed completions are accepted.

The results of the latest completion request are stored and compared
with the next DidChange notification. If the change is at the same
position, more than one character long, and matches a completion
item, the code counts the completion as accepted.

If, for some reason, DidChange returns multiple changes, only the
first is checked.

If the whole file is replaced there is no check for using any
proposed completion.

The test does not exercise the multi-change counter.

Change-Id: I08e9e39f651e4975a86c2dda9f6f5ceab7787e53
Reviewed-on: https://go-review.googlesource.com/c/tools/+/562248
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>