cmd/deadcode: avoid ssautil.AllFunctions
This change causes deadcode to enumerate the "universe set"
of functions based on the source-level declarations, not
the ssautil.AllFunctions helper. The former set is accurate
(and also much cheaper to compute), whereas the latter is
a mess (as its doc comment attests).

In particular, it omits unexported methods of uninstantiated
types, which are of course unreachable. This makes for
some sense for whole-program analysis, but none at all
for a dead code tool that specifically wants to know
the useless functions.

The new logic is pleasantly simpler.

Also, a test.

Fixes golang/go#65915

Change-Id: I7a4a64fd3da18d089044530911cdb2e3b42d0db5
Reviewed-on: https://go-review.googlesource.com/c/tools/+/567156
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>