gopls/internal/golang: improve extract variable edits
The previous implementation use common edits to support both scenarios
where a assignment/declaration exist and did not exist. This cause a
additional text edit with lhs values in the new line when the
extraction was done on a expression which was not part of a
assignment/declaration. The new changes address this scenarios by
computing edits based on the scenario to avoid the additional lhs in
new line when extracted expression is not part of a
assignment/declaration.

Fixes golang/tools#65944