go/callgraph/cha: more precise resolution of unexported methods
Use (*go/types.Func).Id as the key for looking up which methods are
potential matches for a selector.

If a selector refers to an unexported method, it can only be a method
defined in the current package.  Previously the algorithm would
consider all methods of the same name to be potential matches.

SSA functions created by the user with (*ssa.Program).NewFunction may
not have a link to a go/types.Func, and so this would not work, but it
is unlikely that users of the cha library would also use NewFunction to
create a synthetic method on some type.

Fixes golang/go#66689

Change-Id: Ia7d4fcae56987f1d23fbefbf02251d76e5e951c3
Reviewed-on: https://go-review.googlesource.com/c/tools/+/574955
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Tim King <taking@google.com>
Auto-Submit: Alan Donovan <adonovan@google.com>
Reviewed-by: Alan Donovan <adonovan@google.com>