gopls: added method docs to hover info
The change also includes a jump to definition feature for all methods.The
change is quite fail-safe, such that if it fails in any way, it will still
display the hover info without the method docs(as before). The change is
localised to the the gopls/internal/golang/hover.go file, in function
hover and formatHover.
I should also mention that it does not alter the method displayed or it
order, it just add the method docs (if any) and a jump to page source
url to the hover info.

Fixes golang/go#66721
Fixes https://github.com/golang/go/issues/66721

Signed-off-by: tobigiwa <giwaoluwatobi@gmail.com>