internal/imports: a step towards making tests hermetic
If an import test unsets GOROOT, the code looks in the directory /src,
which is usually harmless, but undesirable. That is, it should be
possible for tests to prevent the code from looking anywhere but inside
the constructed test environment. This CL is one step towards that goal.

Change-Id: If14307f653c2b549c5b791a0fbce78c18ce348a4
Reviewed-on: https://go-review.googlesource.com/c/tools/+/579035
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>