go/analysis/passes/nilness: fix bug with MakeInterface(TypeParam)
An interface conversion such as any(x) can be a MakeInterface
(if x is a non-interface type) or a ChangeInterface (if x has
an interface type), and their nilabilities differ.
If x's type is a type parameter, SSA uses MakeInterface, so
we have to ascertain whether the type parameter is definitely
or only maybe a concrete type in order to determine its
nilability.

This change required exposing NormalTerms to x/tools,
and making it take the Underlying of its argument,
so that NormalTerms(error) = NormalTerms(any) = [].
Previously, NormalTerms(error) was [error].

Fixes golang/go#66835
Change-Id: Idf9c39afeaeab918b0f8e6288dd93570f7cb7081
Reviewed-on: https://go-review.googlesource.com/c/tools/+/578938
Reviewed-by: Tim King <taking@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>