gopls/internal/cache: simplify Snapshot Go commands
Address some long-standing TODOs to simplify the invocation of go
commands from the Snapshot. In the past, the combinations of
Snapshot.goCommandInvocation, cache.InvocationFlags, and
Snapshot.RunGoCommand* helpers created a remarkable amount of
indirection even in the case of simple Go command where the arguments
should be known--all on top of the already significant indirection of
the gocommand package. As a result, it was difficult to reason about
what exactly is being run.

This indirection attempted to abstract things like the go command
environment, and -mod flags, which in the past were more configurable
and depended on the ambient Go version. But we've since stopped
supporting much of this configuration, and the Go command has stabilized
its behavior. The abstraction is no longer carrying its weight.

Simplify as follows:
- Reduce the gopls APIs to (1) Snapshot.GoCommandInvocation, which
  populates an invocation according to the Snapshot's environment, and
  (2) View.GoCommandRunner(), which is used for running Go commands in a
  shared control plane.
- Eliminate cache.InvocationFlags, which were a very leaky abstraction
  around -mod, -modfile, and GOWORK behavior. Instead, pass -mod and
  -modfile as needed at the callsites. In most contexts, it is clear
  what the flags and environment should be, the one notable exception
  being Snapshot.RunGoModUpdateCommands, which may accept multiple
  verbs. This will be cleaned up in a subsequent CL.
- Add a cache.TempModDir helper, to replace the tempURI behavior of
  the WriteTemporaryModFile flag. Now the caller can own the temp
  directory, and snapshot.GoCommandInvocation just produces an
  Invocation.

Change-Id: Ieb5a4dc6e79ce6e69a0f1e072843538f0162f744
Reviewed-on: https://go-review.googlesource.com/c/tools/+/579439
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>