gopls/internal/cache: fix bug.Report converting Diagnostic positions
Analyzers are free to add new files to Pass.FileSet.
(This is documented, but perhaps not clearly enough,
so this CL clarifies the commentary.)

Previously, gopls' analysis driver assumed that the token.File
for a given Diagnostic must be one of the set of parsed Go files
fed to the Pass; the previous point refutes this assumption.
This change causes the driver to search for files by name,
not *token.File pointer, to allow for re-parsing.
The driver still calls bug.Report if the file is not found;
this still indicates a likely analyzer bug.
This should stem the prolific flow of telemetry field reports
from this cause.

Also, add a regression test that ensures that the cgocall analyzer,
which not only parses but type-checks files anew, actually
reports useful diagnostics when run from within gopls.
(Previously, it would trigger the bug.Report.)

Also, plumb the Snapshot's ReadFile and context down to
the Pass so that Pass.ReadFile can make consistent reads
from the snapshot instead of using os.ReadFile directly.
We also reject attempts by the analyzer to read files
other than the permitted ones.

Fixes golang/go#66911
Fixes golang/go#64547

Change-Id: I5c35ad6cc5c15c99e9af48aaffb3df2aff621968
Reviewed-on: https://go-review.googlesource.com/c/tools/+/580836
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>