gopls/internal/cache: support overlays
This change causes go commands executed by the snapshot
to set the -overlay=... flag to a file describing
the unsaved editor buffers of the snapshot. (Previously,
this was done only for the main metadata load via
packages.Load.)

We factor the WriteOverlays logic from go/packages/golist.go
into internal/gocommand and use it from Snapshot.goCommandInvocation.
The cleanup logic has been cleaned up.

The ToggleGCDetails feature was an example of a command
that had a needless "file save required" restriction even
though the underlying machinery (go build) is overlay-aware.
This change removes that restriction and adds a test that
the feature works on unsaved editor buffers.

Another one is Tidy, for which I have also removed the
restriction and added a test.

(This is what happens when you try to write down a list of
all gopls features for documentation purposes...)

Change-Id: I801e3a9c7c27f6b63efaaa1257fcca37e6fafa4c
Reviewed-on: https://go-review.googlesource.com/c/tools/+/582937
Auto-Submit: Alan Donovan <adonovan@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>