gopls/internal/protocol: rename DocumentChange{s,}
DocumentChange is a singular thing: a sum of four cases.
It is the element type of the WorkspaceEdit.DocumentChanges
slice field.

(The formatting of this type seen in
https://github.com/golang/tools/blob/gopls-release-branch.0.15/gopls/doc/commands.md#apply-a-fix
and commented on in CL 585277 is a consequence of structDoc
in gopls/doc/generate/generate.go; it remains unchanged.)

Change-Id: Idcd766a8a9c3228e5c43929fbf5dd6795ee7d301
Reviewed-on: https://go-review.googlesource.com/c/tools/+/585975
Reviewed-by: Robert Findley <rfindley@google.com>
Auto-Submit: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>