internal/test/marker: skip basiclit.txt on ppc64
This test was flaking in a bizarre manner on the aix-ppc64 builder. The
x/text/unicode/runenames package apparently returns a different name for
U+2211.

Since aix-ppc64 is not a first class port, skip to de-flake. However,
the failure mode is so bizarre that I added a unit test in the golang
package to try to reproduce the bug using only the x/text module. If
this new unit test also flakes, I will file a bug against aix-ppc64 and
x/text.

Fixes golang/go#65072

Change-Id: I99a6b60c7fd31b474e0b670e6f26e550de176ba8
Reviewed-on: https://go-review.googlesource.com/c/tools/+/586260
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Peter Weinberger <pjw@google.com>