go/analysis/passes/copylock: suppress error in ill-typed code
The copylock analyzer is marked RunDespiteErrors. In ill-typed
code such as S{T} where S and T are both types, the compiler will
warn that T is not an expression; the copylocks analyzer should
not additionally report a diagnostic as if T had been an expression
of type T.

The fix feels rather ad hoc. In general, analyzers marked
RunDespiteErrors are unlikely to be able to anticipate the
myriad ways that trees can be ill-formed and avoid spurious
diagnostics.

Also, add a main.go file for copylock.

Fixes golang/go#67787

Change-Id: I07afbed16a4138fe602c22ec42171b4a5e634286
Reviewed-on: https://go-review.googlesource.com/c/tools/+/589895
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>