internal/stdlib: generate from api/next/*.txt (go1.23) too
This CL causes the generator to treat the next/*.txt files
as if they comprise the next (e.g. go1.23.txt) manifest,
which doesn't exist until close to the release (specifically,
between the start of the freeze and the first release
candidate).

Change-Id: I84158032484d5aa4b1d2d57c11209a24e6924b5e
Reviewed-on: https://go-review.googlesource.com/c/tools/+/590355
Reviewed-by: Russ Cox <rsc@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>