gopls/internal/test: add test case for parameter rename match import
This adds a test case to verify that valid renamings can occur, when
the new name of a parameter is identical to an imported package, even
when referenced by a function paramater.

The bug where the rename was failing is fixed when using go1.22. The
new go version fills in a missing Pos for an object scope that allows
LookupPos to correctly determine if the scope is relevant.

For golang/go#57479

Change-Id: Ifebafb78483f174eac94cd1ec5ac68db9e88684f
Reviewed-on: https://go-review.googlesource.com/c/tools/+/590976
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Alan Donovan <adonovan@google.com>